Integrate intervals of day chart with new backend endpoint.
