+++ frontend/src/reducers/fetchGraphReducer.js
@@ -1,6 +1,7 @@
 const initialState = {
   fetching: false,
   graphData: null,
+  intervalData: null,
 };
 
 export default (state = initialState, action) => {
@@ -11,6 +12,14 @@ export default (state = initialState, action) => {
       return {...state, fetched:false, err:null, graphData:null};
     case "RECEIVED_GRAPH_ERROR":
       return {...state, err: action.payload, graphData:null};
+      
+    case "RECEIVED_INTERVAL_DATA":
+        return {...state, fetched:true, err:null, intervalData:action.payload};
+    case "RESET_INTERVAL_DATA":
+        return {...state, fetched:false, err:null, intervalData:null};
+    case "RECEIVED_INTERVAL_ERROR":
+        return {...state, err: action.payload, intervalData:null};
+        
     default:
       break;
   }
